working on formatting
