+++ src/Illuminate/Database/Query/Builder.php
@@ -3,7 +3,6 @@
 namespace Illuminate\Database\Query;
 
 use BackedEnum;
-use Carbon\Carbon;
 use Carbon\CarbonPeriod;
 use Closure;
 use DateTimeInterface;
@@ -11,6 +10,7 @@ use Illuminate\Contracts\Database\Query\Builder as BuilderContract;
 use Illuminate\Contracts\Database\Query\ConditionExpression;
 use Illuminate\Contracts\Database\Query\Expression as ExpressionContract;
 use Illuminate\Contracts\Support\Arrayable;
+use Illuminate\Database\Concerns\BuildsDateWhereClauses;
 use Illuminate\Database\Concerns\BuildsQueries;
 use Illuminate\Database\Concerns\ExplainsQueries;
 use Illuminate\Database\ConnectionInterface;
@@ -35,7 +35,7 @@ use function Illuminate\Support\enum_value;
 class Builder implements BuilderContract
 {
     /** @use \Illuminate\Database\Concerns\BuildsQueries<object> */
-    use BuildsQueries, ExplainsQueries, ForwardsCalls, Macroable {
+    use BuildsDateWhereClauses, BuildsQueries, ExplainsQueries, ForwardsCalls, Macroable {
         __call as macroCall;
     }
 
@@ -252,9 +252,6 @@ class Builder implements BuilderContract
     /**
      * Create a new query builder instance.
      *
-     * @param  \Illuminate\Database\ConnectionInterface  $connection
-     * @param  \Illuminate\Database\Query\Grammars\Grammar|null  $grammar
-     * @param  \Illuminate\Database\Query\Processors\Processor|null  $processor
      * @return void
      */
     public function __construct(
@@ -313,7 +310,6 @@ class Builder implements BuilderContract
      * Add a new "raw" select expression to the query.
      *
      * @param  string  $expression
-     * @param  array  $bindings
      * @return $this
      */
     public function selectRaw($expression, array $bindings = [])
@@ -610,8 +606,6 @@ class Builder implements BuilderContract
      * Add a lateral join clause to the query.
      *
      * @param  \Closure|\Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder<*>|string  $query
-     * @param  string  $as
-     * @param  string  $type
      * @return $this
      */
     public function joinLateral($query, string $as, string $type = 'inner')
@@ -631,7 +625,6 @@ class Builder implements BuilderContract
      * Add a lateral left join to the query.
      *
      * @param  \Closure|\Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder<*>|string  $query
-     * @param  string  $as
      * @return $this
      */
     public function leftJoinLateral($query, string $as)
@@ -768,7 +761,6 @@ class Builder implements BuilderContract
     /**
      * Get a new join clause.
      *
-     * @param  \Illuminate\Database\Query\Builder  $parentQuery
      * @param  string  $type
      * @param  \Illuminate\Contracts\Database\Query\Expression|string  $table
      * @return \Illuminate\Database\Query\JoinClause
@@ -781,7 +773,6 @@ class Builder implements BuilderContract
     /**
      * Get a new join lateral clause.
      *
-     * @param  \Illuminate\Database\Query\Builder  $parentQuery
      * @param  string  $type
      * @param  \Illuminate\Contracts\Database\Query\Expression|string  $table
      * @return \Illuminate\Database\Query\JoinLateralClause
@@ -1383,7 +1374,6 @@ class Builder implements BuilderContract
      * Add a where between statement to the query.
      *
      * @param  \Illuminate\Contracts\Database\Query\Expression|string  $column
-     * @param  iterable  $values
      * @param  string  $boolean
      * @param  bool  $not
      * @return $this
@@ -1407,7 +1397,6 @@ class Builder implements BuilderContract
      * Add a where between statement using columns to the query.
      *
      * @param  \Illuminate\Contracts\Database\Query\Expression|string  $column
-     * @param  array  $values
      * @param  string  $boolean
      * @param  bool  $not
      * @return $this
@@ -1425,7 +1414,6 @@ class Builder implements BuilderContract
      * Add an or where between statement to the query.
      *
      * @param  \Illuminate\Contracts\Database\Query\Expression|string  $column
-     * @param  iterable  $values
      * @return $this
      */
     public function orWhereBetween($column, iterable $values)
@@ -1437,7 +1425,6 @@ class Builder implements BuilderContract
      * Add an or where between statement using columns to the query.
      *
      * @param  \Illuminate\Contracts\Database\Query\Expression|string  $column
-     * @param  array  $values
      * @return $this
      */
     public function orWhereBetweenColumns($column, array $values)
@@ -1449,7 +1436,6 @@ class Builder implements BuilderContract
      * Add a where not between statement to the query.
      *
      * @param  \Illuminate\Contracts\Database\Query\Expression|string  $column
-     * @param  iterable  $values
      * @param  string  $boolean
      * @return $this
      */
@@ -1462,7 +1448,6 @@ class Builder implements BuilderContract
      * Add a where not between statement using columns to the query.
      *
      * @param  \Illuminate\Contracts\Database\Query\Expression|string  $column
-     * @param  array  $values
      * @param  string  $boolean
      * @return $this
      */
@@ -1475,7 +1460,6 @@ class Builder implements BuilderContract
      * Add an or where not between statement to the query.
      *
      * @param  \Illuminate\Contracts\Database\Query\Expression|string  $column
-     * @param  iterable  $values
      * @return $this
      */
     public function orWhereNotBetween($column, iterable $values)
@@ -1487,7 +1471,6 @@ class Builder implements BuilderContract
      * Add an or where not between statement using columns to the query.
      *
      * @param  \Illuminate\Contracts\Database\Query\Expression|string  $column
-     * @param  array  $values
      * @return $this
      */
     public function orWhereNotBetweenColumns($column, array $values)
@@ -1504,180 +1487,6 @@ class Builder implements BuilderContract
     public function orWhereNotNull($column)
     {
         return $this->whereNotNull($column, 'or');
-    }
-
-    /**
-     * Add a where clause to determine if a "date" column is in the past to the query.
-     *
-     * @param  array|string  $columns
-     * @param  \DateTimeInterface|string|null  $now
-     * @param  string  $boolean
-     * @param  bool  $not
-     * @return $this
-     */
-    public function wherePast($columns, $value = null, $boolean = 'and', $not = false)
-    {
-        $type = 'Basic';
-        $operator = $not ? '>=' : '<';
-        $value = $value ?? Carbon::now();
-
-        foreach (Arr::wrap($columns) as $column) {
-            $this->wheres[] = compact('type', 'column', 'boolean', 'operator', 'value');
-
-            $this->addBinding($value);
-        }
-
-        return $this;
-    }
-
-    /**
-     * Add an "or where" clause to determine if a "date" column is in the past to the query.
-     *
-     * @param  array|string  $columns
-     * @param  \DateTimeInterface|string|null  $value
-     * @return $this
-     */
-    public function orWherePast($columns, $value = null)
-    {
-        return $this->wherePast($columns, $value, 'or');
-    }
-
-    /**
-     * Add a where clause to determine if a "date" column is not in the past to the query.
-     *
-     * @param  array|string  $columns
-     * @param  \DateTimeInterface|string|null  $value
-     * @return $this
-     */
-    public function whereNotPast($columns, $value = null)
-    {
-        return $this->wherePast($columns, $value, 'and', true);
-    }
-
-    /**
-     * Add an "or where" clause to determine if a "date" column is in the past to the query.
-     *
-     * @param  array|string  $columns
-     * @param  \DateTimeInterface|string|null  $value
-     * @return $this
-     */
-    public function orWhereNotPast($columns, $value = null)
-    {
-        return $this->wherePast($columns, $value, 'or', true);
-    }
-
-    /**
-     * Add a "where date" clause to determine if a "date" column is today to the query.
-     *
-     * @param  array|string  $columns
-     * @param  string  $boolean
-     * @param  bool  $not
-     * @return $this
-     */
-    public function whereToday($columns, $boolean = 'and', $not = false)
-    {
-        $operator = $not ? '!=' : '=';
-
-        $value = Carbon::now()->format('Y-m-d');
-
-        foreach (Arr::wrap($columns) as $column) {
-            $this->addDateBasedWhere('Date', $column, $operator, $value, $boolean);
-        }
-
-        return $this;
-    }
-
-    /**
-     * Add an "or where date" clause to determine if a "date" column is today to the query.
-     *
-     * @param  array|string  $columns
-     * @return $this
-     */
-    public function orWhereToday($columns)
-    {
-        return $this->whereToday($columns, 'or');
-    }
-
-    /**
-     * Add a "where date" clause to determine if a "date" column is not today to the query.
-     *
-     * @param  array|string  $columns
-     * @return $this
-     */
-    public function whereNotToday($columns)
-    {
-        return $this->whereToday($columns, 'and', true);
-    }
-
-    /**
-     * Add an "or where date" clause to determine if a "date" column is not today to the query.
-     *
-     * @param  array|string  $columns
-     * @return $this
-     */
-    public function orWhereNotToday($columns)
-    {
-        return $this->whereToday($columns, 'or', true);
-    }
-
-    /**
-     * Add a where clause to determine if a "date" column is in the future to the query.
-     *
-     * @param  array|string  $columns
-     * @param  \DateTimeInterface|string|null  $value
-     * @param  string  $boolean
-     * @param  bool  $not
-     * @return $this
-     */
-    public function whereFuture($columns, $value = null, $boolean = 'and', $not = false)
-    {
-        $type = 'Basic';
-        $operator = $not ? '<=' : '>';
-        $value = $value ?? Carbon::now();
-
-        foreach (Arr::wrap($columns) as $column) {
-            $this->wheres[] = compact('type', 'column', 'boolean', 'operator', 'value');
-
-            $this->addBinding($value);
-        }
-
-        return $this;
-    }
-
-    /**
-     * Add an "or where" clause to determine if a "date" column is in the future to the query.
-     *
-     * @param  array|string  $columns
-     * @param  \DateTimeInterface|string|null  $value
-     * @return $this
-     */
-    public function orWhereFuture($columns, $value = null)
-    {
-        return $this->whereFuture($columns, $value, 'or');
-    }
-
-    /**
-     * Add a where clause to determine if a "date" column is not in the future to the query.
-     *
-     * @param  array|string  $columns
-     * @param  \DateTimeInterface|string|null  $value
-     * @return $this
-     */
-    public function whereNotFuture($columns, $value = null)
-    {
-        return $this->whereFuture($columns, $value, 'and', true);
-    }
-
-    /**
-     * Add an "or where" clause to determine if a "date" column is in the future to the query.
-     *
-     * @param  array|string  $columns
-     * @param  \DateTimeInterface|string|null  $value
-     * @return $this
-     */
-    public function orWhereNotFuture($columns, $value = null)
-    {
-        return $this->whereFuture($columns, $value, 'or', true);
     }
 
     /**
@@ -1952,7 +1761,6 @@ class Builder implements BuilderContract
     /**
      * Add a nested where statement to the query.
      *
-     * @param  \Closure  $callback
      * @param  string  $boolean
      * @return $this
      */
@@ -2086,7 +1894,6 @@ class Builder implements BuilderContract
     /**
      * Add an exists clause to the query.
      *
-     * @param  \Illuminate\Database\Query\Builder  $query
      * @param  string  $boolean
      * @param  bool  $not
      * @return $this
@@ -2578,7 +2385,6 @@ class Builder implements BuilderContract
      * Add a raw groupBy clause to the query.
      *
      * @param  string  $sql
-     * @param  array  $bindings
      * @return $this
      */
     public function groupByRaw($sql, array $bindings = [])
@@ -2662,7 +2468,6 @@ class Builder implements BuilderContract
     /**
      * Add a nested having statement to the query.
      *
-     * @param  \Closure  $callback
      * @param  string  $boolean
      * @return $this
      */
@@ -2750,7 +2555,6 @@ class Builder implements BuilderContract
      * Add a "having between " clause to the query.
      *
      * @param  string  $column
-     * @param  iterable  $values
      * @param  string  $boolean
      * @param  bool  $not
      * @return $this
@@ -2774,7 +2578,6 @@ class Builder implements BuilderContract
      * Add a raw having clause to the query.
      *
      * @param  string  $sql
-     * @param  array  $bindings
      * @param  string  $boolean
      * @return $this
      */
@@ -2793,7 +2596,6 @@ class Builder implements BuilderContract
      * Add a raw or having clause to the query.
      *
      * @param  string  $sql
-     * @param  array  $bindings
      * @return $this
      */
     public function orHavingRaw($sql, array $bindings = [])
@@ -3125,7 +2927,6 @@ class Builder implements BuilderContract
     /**
      * Register a closure to be invoked before the query is executed.
      *
-     * @param  callable  $callback
      * @return $this
      */
     public function beforeQuery(callable $callback)
@@ -3152,7 +2953,6 @@ class Builder implements BuilderContract
     /**
      * Register a closure to be invoked after the query is executed.
      *
-     * @param  \Closure  $callback
      * @return $this
      */
     public function afterQuery(Closure $callback)
@@ -3254,8 +3054,6 @@ class Builder implements BuilderContract
     /**
      * Get a single expression value from the first result of a query.
      *
-     * @param  string  $expression
-     * @param  array  $bindings
      * @return mixed
      */
     public function rawValue(string $expression, array $bindings = [])
@@ -3501,7 +3299,6 @@ class Builder implements BuilderContract
     /**
      * Remove the column aliases since they will break count queries.
      *
-     * @param  array  $columns
      * @return array
      */
     protected function withoutSelectAliases(array $columns)
@@ -3706,7 +3503,6 @@ class Builder implements BuilderContract
     /**
      * Execute the given callback if no rows exist for the current query.
      *
-     * @param  \Closure  $callback
      * @return mixed
      */
     public function existsOr(Closure $callback)
@@ -3717,7 +3513,6 @@ class Builder implements BuilderContract
     /**
      * Execute the given callback if rows exist for the current query.
      *
-     * @param  \Closure  $callback
      * @return mixed
      */
     public function doesntExistOr(Closure $callback)
@@ -3888,7 +3683,6 @@ class Builder implements BuilderContract
     /**
      * Insert new records into the database.
      *
-     * @param  array  $values
      * @return bool
      */
     public function insert(array $values)
@@ -3929,7 +3723,6 @@ class Builder implements BuilderContract
     /**
      * Insert new records into the database while ignoring errors.
      *
-     * @param  array  $values
      * @return int
      */
     public function insertOrIgnore(array $values)
@@ -3959,7 +3752,6 @@ class Builder implements BuilderContract
     /**
      * Insert a new record and get the value of the primary key.
      *
-     * @param  array  $values
      * @param  string|null  $sequence
      * @return int
      */
@@ -3977,7 +3769,6 @@ class Builder implements BuilderContract
     /**
      * Insert new records into the table using a subquery.
      *
-     * @param  array  $columns
      * @param  \Closure|\Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder<*>|string  $query
      * @return int
      */
@@ -3996,7 +3787,6 @@ class Builder implements BuilderContract
     /**
      * Insert new records into the table using a subquery while ignoring errors.
      *
-     * @param  array  $columns
      * @param  \Closure|\Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder<*>|string  $query
      * @return int
      */
@@ -4015,7 +3805,6 @@ class Builder implements BuilderContract
     /**
      * Update records in the database.
      *
-     * @param  array  $values
      * @return int
      */
     public function update(array $values)
@@ -4046,7 +3835,6 @@ class Builder implements BuilderContract
     /**
      * Update records in a PostgreSQL database using the update from syntax.
      *
-     * @param  array  $values
      * @return int
      */
     public function updateFrom(array $values)
@@ -4067,8 +3855,6 @@ class Builder implements BuilderContract
     /**
      * Insert or update a record matching the attributes, and fill it with values.
      *
-     * @param  array  $attributes
-     * @param  array|callable  $values
      * @return bool
      */
     public function updateOrInsert(array $attributes, array|callable $values = [])
@@ -4093,7 +3879,6 @@ class Builder implements BuilderContract
     /**
      * Insert new records or update the existing ones.
      *
-     * @param  array  $values
      * @param  array|string  $uniqueBy
      * @param  array|null  $update
      * @return int
@@ -4140,7 +3925,6 @@ class Builder implements BuilderContract
      *
      * @param  string  $column
      * @param  float|int  $amount
-     * @param  array  $extra
      * @return int
      *
      * @throws \InvalidArgumentException
@@ -4183,7 +3967,6 @@ class Builder implements BuilderContract
      *
      * @param  string  $column
      * @param  float|int  $amount
-     * @param  array  $extra
      * @return int
      *
      * @throws \InvalidArgumentException
@@ -4337,7 +4120,6 @@ class Builder implements BuilderContract
     /**
      * Set the bindings on the query builder.
      *
-     * @param  array  $bindings
      * @param  string  $type
      * @return $this
      *
@@ -4399,7 +4181,6 @@ class Builder implements BuilderContract
     /**
      * Merge an array of bindings into our bindings.
      *
-     * @param  \Illuminate\Database\Query\Builder  $query
      * @return $this
      */
     public function mergeBindings(self $query)
@@ -4412,7 +4193,6 @@ class Builder implements BuilderContract
     /**
      * Remove all of the expressions from a list of bindings.
      *
-     * @param  array  $bindings
      * @return array
      */
     public function cleanBindings(array $bindings)
@@ -4516,7 +4296,6 @@ class Builder implements BuilderContract
     /**
      * Clone the query without the given properties.
      *
-     * @param  array  $properties
      * @return static
      */
     public function cloneWithout(array $properties)
@@ -4531,7 +4310,6 @@ class Builder implements BuilderContract
     /**
      * Clone the query without the given bindings.
      *
-     * @param  array  $except
      * @return static
      */
     public function cloneWithoutBindings(array $except)
